train-data-ingestor/requirements.txt
azure-functions
pandas
requests
pymssql
sqlalchemy